+++ google/resource_compute_firewall.go
@@ -8,8 +8,18 @@ import (
 
 	"github.com/hashicorp/terraform/helper/hashcode"
 	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform/helper/validation"
+
+	computeBeta "google.golang.org/api/compute/v0.beta"
 	"google.golang.org/api/compute/v1"
 )
+
+var FirewallBaseApiVersion = v1
+var FirewallVersionedFeatures = []Feature{
+	Feature{Version: v0beta, Item: "deny"},
+	Feature{Version: v0beta, Item: "direction"},
+	Feature{Version: v0beta, Item: "destination_ranges"},
+}
 
 func resourceComputeFirewall() *schema.Resource {
 	return &schema.Resource{
@@ -37,8 +47,9 @@ func resourceComputeFirewall() *schema.Resource {
 			},
 
 			"allow": {
-				Type:     schema.TypeSet,
-				Required: true,
+				Type:          schema.TypeSet,
+				Optional:      true,
+				ConflictsWith: []string{"deny"},
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"protocol": {
@@ -53,12 +64,45 @@ func resourceComputeFirewall() *schema.Resource {
 						},
 					},
 				},
-				Set: resourceComputeFirewallAllowHash,
+				Set: resourceComputeFirewallRuleHash,
+			},
+
+			"deny": {
+				Type:          schema.TypeSet,
+				Optional:      true,
+				ConflictsWith: []string{"allow"},
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"protocol": {
+							Type:     schema.TypeString,
+							Required: true,
+							ForceNew: true,
+						},
+
+						"ports": {
+							Type:     schema.TypeList,
+							Optional: true,
+							Elem:     &schema.Schema{Type: schema.TypeString},
+							ForceNew: true,
+						},
+					},
+				},
+				Set: resourceComputeFirewallRuleHash,
+
+				// Unlike allow, deny can't be updated upstream
+				ForceNew: true,
 			},
 
 			"description": {
 				Type:     schema.TypeString,
 				Optional: true,
+			},
+
+			"direction": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				ValidateFunc: validation.StringInSlice([]string{"INGRESS", "EGRESS"}, false),
+				ForceNew:     true,
 			},
 
 			"project": {
@@ -88,6 +132,16 @@ func resourceComputeFirewall() *schema.Resource {
 				Set:      schema.HashString,
 			},
 
+			"destination_ranges": {
+				Type:          schema.TypeSet,
+				Optional:      true,
+				Computed:      true,
+				ConflictsWith: []string{"source_ranges", "source_tags"},
+				Elem:          &schema.Schema{Type: schema.TypeString},
+				Set:           schema.HashString,
+				ForceNew:      true,
+			},
+
 			"target_tags": {
 				Type:     schema.TypeSet,
 				Optional: true,
@@ -98,7 +152,7 @@ func resourceComputeFirewall() *schema.Resource {
 	}
 }
 
-func resourceComputeFirewallAllowHash(v interface{}) int {
+func resourceComputeFirewallRuleHash(v interface{}) int {
 	var buf bytes.Buffer
 	m := v.(map[string]interface{})
 	buf.WriteString(fmt.Sprintf("%s-", m["protocol"].(string)))
@@ -118,6 +172,7 @@ func resourceComputeFirewallAllowHash(v interface{}) int {
 }
 
 func resourceComputeFirewallCreate(d *schema.ResourceData, meta interface{}) error {
+	computeApiVersion := getComputeApiVersion(d, FirewallBaseApiVersion, FirewallVersionedFeatures)
 	config := meta.(*Config)
 
 	project, err := getProject(d, config)
@@ -125,21 +180,41 @@ func resourceComputeFirewallCreate(d *schema.ResourceData, meta interface{}) err
 		return err
 	}
 
-	firewall, err := resourceFirewall(d, meta)
+	firewall, err := resourceFirewall(d, meta, computeApiVersion)
 	if err != nil {
 		return err
 	}
 
-	op, err := config.clientCompute.Firewalls.Insert(
-		project, firewall).Do()
-	if err != nil {
-		return fmt.Errorf("Error creating firewall: %s", err)
+	var op interface{}
+	switch computeApiVersion {
+	case v1:
+		firewallV1 := &compute.Firewall{}
+		err = Convert(firewall, firewallV1)
+		if err != nil {
+			return err
+		}
+
+		op, err = config.clientCompute.Firewalls.Insert(project, firewallV1).Do()
+		if err != nil {
+			return fmt.Errorf("Error creating firewall: %s", err)
+		}
+	case v0beta:
+		firewallV0Beta := &computeBeta.Firewall{}
+		err = Convert(firewall, firewallV0Beta)
+		if err != nil {
+			return err
+		}
+
+		op, err = config.clientComputeBeta.Firewalls.Insert(project, firewallV0Beta).Do()
+		if err != nil {
+			return fmt.Errorf("Error creating firewall: %s", err)
+		}
 	}
 
 	// It probably maybe worked, so store the ID now
 	d.SetId(firewall.Name)
 
-	err = computeOperationWait(config, op, project, "Creating Firewall")
+	err = computeSharedOperationWait(config, op, project, "Creating Firewall")
 	if err != nil {
 		return err
 	}
@@ -147,7 +222,7 @@ func resourceComputeFirewallCreate(d *schema.ResourceData, meta interface{}) err
 	return resourceComputeFirewallRead(d, meta)
 }
 
-func flattenAllowed(allowed []*compute.FirewallAllowed) []map[string]interface{} {
+func flattenAllowed(allowed []*computeBeta.FirewallAllowed) []map[string]interface{} {
 	result := make([]map[string]interface{}, 0, len(allowed))
 	for _, allow := range allowed {
 		allowMap := make(map[string]interface{})
@@ -159,7 +234,20 @@ func flattenAllowed(allowed []*compute.FirewallAllowed) []map[string]interface{}
 	return result
 }
 
+func flattenDenied(denied []*computeBeta.FirewallDenied) []map[string]interface{} {
+	result := make([]map[string]interface{}, 0, len(denied))
+	for _, deny := range denied {
+		denyMap := make(map[string]interface{})
+		denyMap["protocol"] = deny.IPProtocol
+		denyMap["ports"] = deny.Ports
+
+		result = append(result, denyMap)
+	}
+	return result
+}
+
 func resourceComputeFirewallRead(d *schema.ResourceData, meta interface{}) error {
+	computeApiVersion := getComputeApiVersion(d, FirewallBaseApiVersion, FirewallVersionedFeatures)
 	config := meta.(*Config)
 
 	project, err := getProject(d, config)
@@ -167,26 +255,55 @@ func resourceComputeFirewallRead(d *schema.ResourceData, meta interface{}) error
 		return err
 	}
 
-	firewall, err := config.clientCompute.Firewalls.Get(
-		project, d.Id()).Do()
-	if err != nil {
-		return handleNotFoundError(err, d, fmt.Sprintf("Firewall %q", d.Get("name").(string)))
+	firewall := &computeBeta.Firewall{}
+	switch computeApiVersion {
+	case v1:
+		firewallV1, err := config.clientCompute.Firewalls.Get(project, d.Id()).Do()
+		if err != nil {
+			return handleNotFoundError(err, d, fmt.Sprintf("Firewall %q", d.Get("name").(string)))
+		}
+
+		err = Convert(firewallV1, firewall)
+		if err != nil {
+			return err
+		}
+	case v0beta:
+		firewallV0Beta, err := config.clientComputeBeta.Firewalls.Get(project, d.Id()).Do()
+		if err != nil {
+			return handleNotFoundError(err, d, fmt.Sprintf("Firewall %q", d.Get("name").(string)))
+		}
+
+		err = Convert(firewallV0Beta, firewall)
+		if err != nil {
+			return err
+		}
 	}
 
 	networkUrl := strings.Split(firewall.Network, "/")
-	d.Set("self_link", firewall.SelfLink)
+	d.Set("self_link", ConvertSelfLinkToV1(firewall.SelfLink))
 	d.Set("name", firewall.Name)
 	d.Set("network", networkUrl[len(networkUrl)-1])
+
+	// Unlike most other Beta properties, direction will always have a value even when
+	// a zero is sent by the client. We'll never revert back to v1 without conditionally reading it.
+	// This if statement blocks Beta import for this resource.
+	if _, ok := d.GetOk("direction"); ok {
+		d.Set("direction", firewall.Direction)
+	}
+
 	d.Set("description", firewall.Description)
 	d.Set("project", project)
 	d.Set("source_ranges", firewall.SourceRanges)
 	d.Set("source_tags", firewall.SourceTags)
+	d.Set("destination_ranges", firewall.DestinationRanges)
 	d.Set("target_tags", firewall.TargetTags)
 	d.Set("allow", flattenAllowed(firewall.Allowed))
+	d.Set("deny", flattenDenied(firewall.Denied))
 	return nil
 }
 
 func resourceComputeFirewallUpdate(d *schema.ResourceData, meta interface{}) error {
+	computeApiVersion := getComputeApiVersionUpdate(d, FirewallBaseApiVersion, FirewallVersionedFeatures, []Feature{})
 	config := meta.(*Config)
 
 	project, err := getProject(d, config)
@@ -196,18 +313,38 @@ func resourceComputeFirewallUpdate(d *schema.ResourceData, meta interface{}) err
 
 	d.Partial(true)
 
-	firewall, err := resourceFirewall(d, meta)
+	firewall, err := resourceFirewall(d, meta, computeApiVersion)
 	if err != nil {
 		return err
 	}
 
-	op, err := config.clientCompute.Firewalls.Update(
-		project, d.Id(), firewall).Do()
-	if err != nil {
-		return fmt.Errorf("Error updating firewall: %s", err)
+	var op interface{}
+	switch computeApiVersion {
+	case v1:
+		firewallV1 := &compute.Firewall{}
+		err = Convert(firewall, firewallV1)
+		if err != nil {
+			return err
+		}
+
+		op, err = config.clientCompute.Firewalls.Update(project, d.Id(), firewallV1).Do()
+		if err != nil {
+			return fmt.Errorf("Error updating firewall: %s", err)
+		}
+	case v0beta:
+		firewallV0Beta := &computeBeta.Firewall{}
+		err = Convert(firewall, firewallV0Beta)
+		if err != nil {
+			return err
+		}
+
+		op, err = config.clientComputeBeta.Firewalls.Update(project, d.Id(), firewallV0Beta).Do()
+		if err != nil {
+			return fmt.Errorf("Error updating firewall: %s", err)
+		}
 	}
 
-	err = computeOperationWait(config, op, project, "Updating Firewall")
+	err = computeSharedOperationWait(config, op, project, "Updating Firewall")
 	if err != nil {
 		return err
 	}
@@ -218,6 +355,7 @@ func resourceComputeFirewallUpdate(d *schema.ResourceData, meta interface{}) err
 }
 
 func resourceComputeFirewallDelete(d *schema.ResourceData, meta interface{}) error {
+	computeApiVersion := getComputeApiVersion(d, FirewallBaseApiVersion, FirewallVersionedFeatures)
 	config := meta.(*Config)
 
 	project, err := getProject(d, config)
@@ -226,13 +364,21 @@ func resourceComputeFirewallDelete(d *schema.ResourceData, meta interface{}) err
 	}
 
 	// Delete the firewall
-	op, err := config.clientCompute.Firewalls.Delete(
-		project, d.Id()).Do()
-	if err != nil {
-		return fmt.Errorf("Error deleting firewall: %s", err)
+	var op interface{}
+	switch computeApiVersion {
+	case v1:
+		op, err = config.clientCompute.Firewalls.Delete(project, d.Id()).Do()
+		if err != nil {
+			return fmt.Errorf("Error deleting firewall: %s", err)
+		}
+	case v0beta:
+		op, err = config.clientComputeBeta.Firewalls.Delete(project, d.Id()).Do()
+		if err != nil {
+			return fmt.Errorf("Error deleting firewall: %s", err)
+		}
 	}
 
-	err = computeOperationWait(config, op, project, "Deleting Firewall")
+	err = computeSharedOperationWait(config, op, project, "Deleting Firewall")
 	if err != nil {
 		return err
 	}
@@ -241,24 +387,19 @@ func resourceComputeFirewallDelete(d *schema.ResourceData, meta interface{}) err
 	return nil
 }
 
-func resourceFirewall(
-	d *schema.ResourceData,
-	meta interface{}) (*compute.Firewall, error) {
+func resourceFirewall(d *schema.ResourceData, meta interface{}, computeApiVersion ComputeApiVersion) (*computeBeta.Firewall, error) {
 	config := meta.(*Config)
-
 	project, _ := getProject(d, config)
 
-	// Look up the network to attach the firewall to
-	network, err := config.clientCompute.Networks.Get(
-		project, d.Get("network").(string)).Do()
+	network, err := config.clientCompute.Networks.Get(project, d.Get("network").(string)).Do()
 	if err != nil {
 		return nil, fmt.Errorf("Error reading network: %s", err)
 	}
 
 	// Build up the list of allowed entries
-	var allowed []*compute.FirewallAllowed
+	var allowed []*computeBeta.FirewallAllowed
 	if v := d.Get("allow").(*schema.Set); v.Len() > 0 {
-		allowed = make([]*compute.FirewallAllowed, 0, v.Len())
+		allowed = make([]*computeBeta.FirewallAllowed, 0, v.Len())
 		for _, v := range v.List() {
 			m := v.(map[string]interface{})
 
@@ -270,7 +411,29 @@ func resourceFirewall(
 				}
 			}
 
-			allowed = append(allowed, &compute.FirewallAllowed{
+			allowed = append(allowed, &computeBeta.FirewallAllowed{
+				IPProtocol: m["protocol"].(string),
+				Ports:      ports,
+			})
+		}
+	}
+
+	// Build up the list of denied entries
+	var denied []*computeBeta.FirewallDenied
+	if v := d.Get("deny").(*schema.Set); v.Len() > 0 {
+		denied = make([]*computeBeta.FirewallDenied, 0, v.Len())
+		for _, v := range v.List() {
+			m := v.(map[string]interface{})
+
+			var ports []string
+			if v := convertStringArr(m["ports"].([]interface{})); len(v) > 0 {
+				ports = make([]string, len(v))
+				for i, v := range v {
+					ports[i] = v
+				}
+			}
+
+			denied = append(denied, &computeBeta.FirewallDenied{
 				IPProtocol: m["protocol"].(string),
 				Ports:      ports,
 			})
@@ -292,6 +455,15 @@ func resourceFirewall(
 		}
 	}
 
+	// Build up the list of destinations
+	var destinationRanges []string
+	if v := d.Get("destination_ranges").(*schema.Set); v.Len() > 0 {
+		destinationRanges = make([]string, v.Len())
+		for i, v := range v.List() {
+			destinationRanges[i] = v.(string)
+		}
+	}
+
 	// Build up the list of targets
 	var targetTags []string
 	if v := d.Get("target_tags").(*schema.Set); v.Len() > 0 {
@@ -302,13 +474,16 @@ func resourceFirewall(
 	}
 
 	// Build the firewall parameter
-	return &compute.Firewall{
-		Name:         d.Get("name").(string),
-		Description:  d.Get("description").(string),
-		Network:      network.SelfLink,
-		Allowed:      allowed,
-		SourceRanges: sourceRanges,
-		SourceTags:   sourceTags,
-		TargetTags:   targetTags,
+	return &computeBeta.Firewall{
+		Name:              d.Get("name").(string),
+		Description:       d.Get("description").(string),
+		Direction:         d.Get("direction").(string),
+		Network:           network.SelfLink,
+		Allowed:           allowed,
+		Denied:            denied,
+		SourceRanges:      sourceRanges,
+		SourceTags:        sourceTags,
+		DestinationRanges: destinationRanges,
+		TargetTags:        targetTags,
 	}, nil
 }