+++ google/resource_compute_global_address_test.go
@@ -7,6 +7,8 @@ import (
 	"github.com/hashicorp/terraform/helper/acctest"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
+
+	computeBeta "google.golang.org/api/compute/v0.beta"
 	"google.golang.org/api/compute/v1"
 )
 
@@ -23,6 +25,29 @@ func TestAccComputeGlobalAddress_basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckComputeGlobalAddressExists(
 						"google_compute_global_address.foobar", &addr),
+
+					// implicitly IPV4 - if we don't send an ip_version, we don't get one back even when using Beta apis
+					testAccCheckComputeBetaGlobalAddressIpVersion("google_compute_global_address.foobar", ""),
+				),
+			},
+		},
+	})
+}
+
+func TestAccComputeGlobalAddress_ipv6(t *testing.T) {
+	var addr computeBeta.Address
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckComputeGlobalAddressDestroy,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccComputeGlobalAddress_ipv6,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckComputeBetaGlobalAddressExists(
+						"google_compute_global_address.foobar", &addr),
+					testAccCheckComputeBetaGlobalAddressIpVersion("google_compute_global_address.foobar", "IPV6"),
 				),
 			},
 		},
@@ -76,7 +101,67 @@ func testAccCheckComputeGlobalAddressExists(n string, addr *compute.Address) res
 	}
 }
 
+func testAccCheckComputeBetaGlobalAddressExists(n string, addr *computeBeta.Address) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Not found: %s", n)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("No ID is set")
+		}
+
+		config := testAccProvider.Meta().(*Config)
+
+		found, err := config.clientComputeBeta.GlobalAddresses.Get(config.Project, rs.Primary.ID).Do()
+		if err != nil {
+			return err
+		}
+
+		if found.Name != rs.Primary.ID {
+			return fmt.Errorf("Addr not found")
+		}
+
+		*addr = *found
+
+		return nil
+	}
+}
+
+func testAccCheckComputeBetaGlobalAddressIpVersion(n, version string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Not found: %s", n)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("No ID is set")
+		}
+
+		config := testAccProvider.Meta().(*Config)
+
+		addr, err := config.clientComputeBeta.GlobalAddresses.Get(config.Project, rs.Primary.ID).Do()
+		if err != nil {
+			return err
+		}
+
+		if addr.IpVersion != version {
+			return fmt.Errorf("Expected IP version to be %s, got %s", version, addr.IpVersion)
+		}
+
+		return nil
+	}
+}
+
 var testAccComputeGlobalAddress_basic = fmt.Sprintf(`
 resource "google_compute_global_address" "foobar" {
 	name = "address-test-%s"
+}`, acctest.RandString(10))
+
+var testAccComputeGlobalAddress_ipv6 = fmt.Sprintf(`
+resource "google_compute_global_address" "foobar" {
+	name = "address-test-%s"
+	ip_version = "IPV6"
 }`, acctest.RandString(10))