+++ google/resource_container_node_pool.go
@@ -3,11 +3,12 @@ package google
 import (
 	"fmt"
 	"log"
+	"strings"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform/helper/validation"
 	"google.golang.org/api/container/v1"
-	"google.golang.org/api/googleapi"
 )
 
 func resourceContainerNodePool() *schema.Resource {
@@ -17,6 +18,13 @@ func resourceContainerNodePool() *schema.Resource {
 		Update: resourceContainerNodePoolUpdate,
 		Delete: resourceContainerNodePoolDelete,
 		Exists: resourceContainerNodePoolExists,
+
+		SchemaVersion: 1,
+		MigrateState:  resourceContainerNodePoolMigrateState,
+
+		Importer: &schema.ResourceImporter{
+			State: resourceContainerNodePoolStateImporter,
+		},
 
 		Schema: map[string]*schema.Schema{
 			"project": &schema.Schema{
@@ -66,29 +74,15 @@ func resourceContainerNodePool() *schema.Resource {
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"min_node_count": &schema.Schema{
-							Type:     schema.TypeInt,
-							Required: true,
-							ValidateFunc: func(v interface{}, k string) (ws []string, errors []error) {
-								value := v.(int)
-
-								if value < 1 {
-									errors = append(errors, fmt.Errorf("%q must be >=1", k))
-								}
-								return
-							},
+							Type:         schema.TypeInt,
+							Required:     true,
+							ValidateFunc: validation.IntAtLeast(1),
 						},
 
 						"max_node_count": &schema.Schema{
-							Type:     schema.TypeInt,
-							Required: true,
-							ValidateFunc: func(v interface{}, k string) (ws []string, errors []error) {
-								value := v.(int)
-
-								if value < 1 {
-									errors = append(errors, fmt.Errorf("%q must be >=1", k))
-								}
-								return
-							},
+							Type:         schema.TypeInt,
+							Required:     true,
+							ValidateFunc: validation.IntAtLeast(1),
 						},
 					},
 				},
@@ -124,65 +118,7 @@ func resourceContainerNodePoolCreate(d *schema.ResourceData, meta interface{}) e
 	}
 
 	if v, ok := d.GetOk("node_config"); ok {
-		nodeConfigs := v.([]interface{})
-		nodeConfig := nodeConfigs[0].(map[string]interface{})
-
-		nodePool.Config = &container.NodeConfig{}
-
-		if v, ok = nodeConfig["machine_type"]; ok {
-			nodePool.Config.MachineType = v.(string)
-		}
-
-		if v, ok = nodeConfig["disk_size_gb"]; ok {
-			nodePool.Config.DiskSizeGb = int64(v.(int))
-		}
-
-		if v, ok = nodeConfig["local_ssd_count"]; ok {
-			nodePool.Config.LocalSsdCount = int64(v.(int))
-		}
-
-		if v, ok := nodeConfig["oauth_scopes"]; ok {
-			scopesList := v.([]interface{})
-			scopes := []string{}
-			for _, v := range scopesList {
-				scopes = append(scopes, canonicalizeServiceScope(v.(string)))
-			}
-
-			nodePool.Config.OauthScopes = scopes
-		}
-
-		if v, ok = nodeConfig["service_account"]; ok {
-			nodePool.Config.ServiceAccount = v.(string)
-		}
-
-		if v, ok = nodeConfig["metadata"]; ok {
-			m := make(map[string]string)
-			for k, val := range v.(map[string]interface{}) {
-				m[k] = val.(string)
-			}
-			nodePool.Config.Metadata = m
-		}
-
-		if v, ok = nodeConfig["image_type"]; ok {
-			nodePool.Config.ImageType = v.(string)
-		}
-
-		if v, ok = nodeConfig["labels"]; ok {
-			m := make(map[string]string)
-			for k, val := range v.(map[string]interface{}) {
-				m[k] = val.(string)
-			}
-			nodePool.Config.Labels = m
-		}
-
-		if v, ok := nodeConfig["tags"]; ok {
-			tagsList := v.([]interface{})
-			tags := []string{}
-			for _, v := range tagsList {
-				tags = append(tags, v.(string))
-			}
-			nodePool.Config.Tags = tags
-		}
+		nodePool.Config = expandNodeConfig(v)
 	}
 
 	if v, ok := d.GetOk("autoscaling"); ok {
@@ -213,7 +149,7 @@ func resourceContainerNodePoolCreate(d *schema.ResourceData, meta interface{}) e
 
 	log.Printf("[INFO] GKE NodePool %s has been created", name)
 
-	d.SetId(name)
+	d.SetId(fmt.Sprintf("%s/%s/%s", zone, cluster, name))
 
 	return resourceContainerNodePoolRead(d, meta)
 }
@@ -350,13 +286,24 @@ func resourceContainerNodePoolExists(d *schema.ResourceData, meta interface{}) (
 	_, err = config.clientContainer.Projects.Zones.Clusters.NodePools.Get(
 		project, zone, cluster, name).Do()
 	if err != nil {
-		if gerr, ok := err.(*googleapi.Error); ok && gerr.Code == 404 {
-			log.Printf("[WARN] Removing Container NodePool %q because it's gone", name)
-			// The resource doesn't exist anymore
-			return false, err
+		if err = handleNotFoundError(err, d, fmt.Sprintf("Container NodePool %s", d.Get("name").(string))); err == nil {
+			return false, nil
 		}
 		// There was some other error in reading the resource
 		return true, err
 	}
 	return true, nil
+}
+
+func resourceContainerNodePoolStateImporter(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+	parts := strings.Split(d.Id(), "/")
+	if len(parts) != 3 {
+		return nil, fmt.Errorf("Invalid container cluster specifier. Expecting {zone}/{cluster}/{name}")
+	}
+
+	d.Set("zone", parts[0])
+	d.Set("cluster", parts[1])
+	d.Set("name", parts[2])
+
+	return []*schema.ResourceData{d}, nil
 }